move to ts files and do some reorganizing
