+++ package.json
@@ -9,7 +9,7 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@celo-tools/use-contractkit": "^1.2.0",
+    "@celo-tools/use-contractkit": "^1.3.0",
     "@celo/contractkit": "^1.2.5",
     "@emotion/react": "^11.5.0",
     "@emotion/styled": "^11.3.0",
